Move menu prompt code to its own file.
